reduced the number of var in error handling of pipex and some other clean up
