+++ src/pipex_error_handling_utils.c
@@ -6,7 +6,7 @@
 /*   By: nnourine <nnourine@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/03 14:22:09 by asohrabi          #+#    #+#             */
-/*   Updated: 2024/06/19 11:55:58 by nnourine         ###   ########.fr       */
+/*   Updated: 2024/06/19 16:47:26 by nnourine         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,47 +21,6 @@ int	check_after_token(char *str)
 	return (ft_strchr(str, '|') || ft_strchr(str, '<') || ft_strchr(str, '>')
 		|| ((ft_strchr(str, '<') && *(ft_strchr(str, '<') + 1) == '<'))
 		|| ((ft_strchr(str, '>') && *(ft_strchr(str, '>') + 1) == '>')));
-}
-
-void init_token_part1(char ***token)
-{
-    (*token)[0] = "||";
-    (*token)[1] = "<>";
-    (*token)[2] = "<<<";
-    (*token)[3] = "<<";
-    (*token)[4] = "<";
-    (*token)[5] = ">>";
-    (*token)[6] = ">";
-    (*token)[7] = "|";
-    (*token)[8] = "&&";
-    (*token)[9] = "&";
-    (*token)[10] = "*";
-}
-
-void init_token_part2(char ***token)
-{
-	(*token)[11] = "\\";
-    (*token)[12] = ";";
-    (*token)[13] = "(";
-    (*token)[14] = ")";
-    (*token)[15] = "{";
-    (*token)[16] = "}";
-    (*token)[17] = "[";
-    (*token)[18] = "]";
-    (*token)[19] = ":";
-    (*token)[20] = NULL;
-}
-
-char	**init_token(t_env_pack env_pack)
-{
-	char	**token;
-
-	token = (char **)malloc(21 * sizeof(char *));
-	if (!token)
-		clean_all(env_pack.env, env_pack.original_env, NULL, NULL);
-	init_token_part1(&token);
-	init_token_part2(&token);
-	return (token);
 }
 
 char	*find_token(char *cur, t_env_pack env_pack)
@@ -121,11 +80,11 @@ int is_fd_operator(char *str)
 	return (0);
 }
 
-char	*change_token(char *token, char *cur, int *index, int sq_dq, t_env_pack env_pack)
+char	*change_token(t_error_helper *e, char *cur, int *index, t_env_pack env_pack)
 {
 	char	*new_token;
 
-	if (!sq_dq)
+	if (!((e->sq || e->dq)))
 	{
 		new_token = find_token(cur, env_pack);
 		if (new_token)
@@ -139,7 +98,7 @@ char	*change_token(char *token, char *cur, int *index, int sq_dq, t_env_pack env
 			if (*cur != ' ' && !is_fd_operator(cur))
 				return (NULL);
 			else
-				return (token);
+				return (e->token);
 		}
 	}
 	else
